errwrap package
