+++ errors_test.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 	"io"
 	"reflect"
-	"strconv"
 	"strings"
 	"testing"
 )
@@ -280,62 +279,6 @@ func TestErrorEquality(t *testing.T) {
 	}
 }
 
-type errWalkTest struct {
-	cause error
-	sub   []error
-	v     int
-}
-
-func (e *errWalkTest) Error() string {
-	return strconv.Itoa(e.v)
-}
-
-func (e *errWalkTest) Unwrap() error {
-	return e.cause
-}
-
-func (e *errWalkTest) Errors() []error {
-	return e.sub
-}
-
-func testFind(err error, v int) bool {
-	return WalkDeep(err, func(err error) bool {
-		e := err.(*errWalkTest)
-		return e.v == v
-	})
-}
-
-func TestWalkDeep(t *testing.T) {
-	err := &errWalkTest{
-		sub: []error{
-			&errWalkTest{
-				v:     10,
-				cause: &errWalkTest{v: 11},
-			},
-			&errWalkTest{
-				v:     20,
-				cause: &errWalkTest{v: 21, cause: &errWalkTest{v: 22}},
-			},
-			&errWalkTest{
-				v:     30,
-				cause: &errWalkTest{v: 31},
-			},
-		},
-	}
-
-	if !testFind(err, 11) {
-		t.Errorf("not found in first cause chain")
-	}
-
-	if !testFind(err, 22) {
-		t.Errorf("not found in siblings")
-	}
-
-	if testFind(err, 32) {
-		t.Errorf("found not exists")
-	}
-}
-
 type FindMe struct {
 	a int
 }
@@ -387,33 +330,6 @@ func TestFormatWrapped(t *testing.T) {
 	}
 }
 
-type WrappedInPlace struct {
-	*ErrorWrap
-}
-
-func TestErrorWrapper(t *testing.T) {
-	err := WrappedInPlace{&ErrorWrap{New("underlying")}}
-	if err.Error() != "underlying" {
-		t.Errorf("Error()")
-	}
-	err.WrapError(WrapFn("wrap"))
-	if err.Error() != "wrap: underlying" {
-		t.Errorf("wrap Error()")
-	}
-
-	err.WrapError(WrapfFn("wrapf %d", 1))
-	if s := err.Error(); s != "wrapf 1: wrap: underlying" {
-		t.Errorf("wrapf Error() %s", s)
-	}
-
-	/*
-		err.WrapError(WrapsFn("wraps", "i", 2))
-		if s := err.Error(); s != "wraps i=2: wrapf 1: wrap: underlying" {
-			t.Errorf("wrapf Error() %s", s)
-		}
-	*/
-}
-
 type ErrArray []error
 
 func (ea ErrArray) Error() string {
@@ -442,47 +358,4 @@ func TestIsNil(t *testing.T) {
 	if IsNil(ErrArray([]error{nilError{}})) == true {
 		t.Errorf("IsNil expected false")
 	}
-}
-
-func TestWalkDeepComplexTree(t *testing.T) {
-	err := &errWalkTest{v: 1, cause: &errWalkTest{
-		sub: []error{
-			&errWalkTest{
-				v:     10,
-				cause: &errWalkTest{v: 11},
-			},
-			&errWalkTest{
-				v: 20,
-				sub: []error{
-					&errWalkTest{v: 21},
-					&errWalkTest{v: 22},
-				},
-			},
-			&errWalkTest{
-				v:     30,
-				cause: &errWalkTest{v: 31},
-			},
-		},
-	}}
-
-	assertFind := func(v int, comment string) {
-		if !testFind(err, v) {
-			t.Errorf("%d not found in the error: %s", v, comment)
-		}
-	}
-	assertNotFind := func(v int, comment string) {
-		if testFind(err, v) {
-			t.Errorf("%d found in the error, but not expected: %s", v, comment)
-		}
-	}
-
-	assertFind(1, "shallow search")
-	assertFind(11, "deep search A1")
-	assertFind(21, "deep search A2")
-	assertFind(22, "deep search B1")
-	assertNotFind(23, "deep search Neg")
-	assertFind(31, "deep search B2")
-	assertNotFind(32, "deep search Neg")
-	assertFind(30, "Tree node A")
-	assertFind(20, "Tree node with many children")
 }
