+++ errwrap/go_123.go
@@ -1,6 +1,6 @@
 //go:build go1.23
 
-package errors
+package errwrap
 
 import (
 	"iter"
