+++ example_test.go
@@ -151,30 +151,6 @@ func ExampleAsType() {
 	// Output: empty
 }
 
-type errorGroup struct {
-	errs []error
-}
-
-func (eg *errorGroup) Error() string {
-	return errors.Join(eg.errs...).Error()
-}
-
-func (eg *errorGroup) Errors() []error { return eg.errs }
-
-func ExampleErrors() {
-	var eg errorGroup
-	eg.errs = append(eg.errs, errors.New("error1"))
-	eg.errs = append(eg.errs, errors.New("error2"))
-
-	fmt.Println(errors.Errors(nil))
-	fmt.Println(errors.Errors(errors.New("test")))
-	fmt.Println(errors.Errors(&eg))
-	// Output:
-	// []
-	// []
-	// [error1 error2]
-}
-
 func ExampleIsNil() {
 	var empty *ErrEmpty = nil //nolint:staticcheck
 	var err error = empty
@@ -183,33 +159,4 @@ func ExampleIsNil() {
 	// Output:
 	// false
 	// true
-}
-
-type inplace struct {
-	*errors.ErrorWrap
-}
-
-func ExampleWrapInPlace() {
-	err := inplace{errors.NewErrorWrap(errors.New("original error"))}
-
-	// Wrap the error in place
-	wrapped := errors.WrapInPlace(err, errors.WrapFn("wrapped"))
-
-	// Print the error and whether it was wrapped in place
-	fmt.Printf("Wrapped in place: %v\n", wrapped)
-	fmt.Printf("Error: %v\n", err)
-
-	// Try with a regular error that doesn't implement ErrorWrapper
-	regularErr := errors.New("regular error")
-	wrapped = errors.WrapInPlace(regularErr, errors.WrapFn("wrapped"))
-
-	// Print the result for regular error
-	fmt.Printf("Regular error wrapped in place: %v\n", wrapped)
-	fmt.Printf("Regular error: %v\n", regularErr)
-
-	// Output:
-	// Wrapped in place: true
-	// Error: wrapped: original error
-	// Regular error wrapped in place: false
-	// Regular error: regular error
 }
